requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==3.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
Markdown==3.4.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2023.3
PyYAML==6.0
reportlab==3.6.12
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
